+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0] - 2026-01-16
+
 ## [0.8.0] - 2025-11-20
 
 ### Added
@@ -127,7 +129,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] - 2024-08-13
 
-[Unreleased]: https://github.com/giantswarm/cluster-api-events/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/giantswarm/cluster-api-events/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/giantswarm/cluster-api-events/compare/v0.8.0...v1.0.0
 [0.8.0]: https://github.com/giantswarm/cluster-api-events/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/giantswarm/cluster-api-events/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/giantswarm/cluster-api-events/compare/v0.5.3...v0.6.0